refactor: adjust pagination styles for better visibility
